+++ pkg/domain/repository/tourist_spot.go
@@ -12,7 +12,7 @@ type (
 		Lock(c context.Context, id int) (*entity.TouristSpot, error)
 		Store(c context.Context, touristSpot *entity.TouristSpot) error
 		// レビューの平均値を更新
-		UpdateScoreByID(c context.Context, id int) error
+		UpdateScoreByID(c context.Context, id int64) error
 		DeleteByID(id int) error
 	}
